More logical corrections
Rule name Split was getting improper name because stripping white-space was only occurring in conditional, not prior to the split.

Added the ability to return Action Results in a Dict key-ed by the Rule Name as there does not appear to be another way to alter none global objects in the actions other than returning changes.  This could also be useful for storing and returning change messages instead of having to print-stream changes in the action functions.  It also seemed consistent with how the Rule class was working as it returns a value, which is strangely just discarded.  Why return if you're just going to throw it away?